Organize depmap-compute slice utils
And define depmap-compute dependencies
